Bring back ReadAsync/WriteAsync code to the strategy for simpler code review. Shorter filename for awaitable provider.
